requirements.txt
streamlit
openai
langchain
python-dotenv
cassio
